вывел необходимые классы (папка repository, api) в модуль repository.
