+++ app/build.gradle
@@ -31,6 +31,8 @@ android {
 dependencies {
     implementation project(Modules.utils)
     implementation project(Modules.model)
+    implementation project(Modules.repository)
+
 
 // AndroidX
     implementation Design.appcompat
