--- app/src/main/java/com/bignerdranch/android/translator/model/data/AppState.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.android.translator.model.data
-
-sealed class AppState {
-
-    data class Success(val data: List<DataModel>?) : AppState()
-    data class Error(val error: Throwable) : AppState()
-    data class Loading(val progress: Int?) : AppState()
-}
