--- app/src/main/java/com/bignerdranch/android/translator/model/data/DataModel.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.android.translator.model.data
-
-import com.google.gson.annotations.SerializedName
-
-class DataModel(
-    @field:SerializedName("text") val text: String?,
-    @field:SerializedName("meanings") val meanings: List<Meanings>?
-)
