--- app/src/main/java/com/bignerdranch/android/translator/model/data/Meanings.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.android.translator.model.data
-
-import com.google.gson.annotations.SerializedName
-
-class Meanings(
-    @field:SerializedName("translation") val translation: Translation?,
-    @field:SerializedName("imageUrl") val imageUrl: String?
-)
