--- app/src/main/java/com/bignerdranch/android/translator/model/data/Translation.kt
@@ -1,5 +1,0 @@
-package com.bignerdranch.android.translator.model.data
-
-import com.google.gson.annotations.SerializedName
-
-class Translation(@field:SerializedName("text") val translation: String?)
