--- app/src/main/java/com/bignerdranch/android/translator/model/data/api/ApiService.kt
@@ -1,12 +1,0 @@
-package com.bignerdranch.android.translator.model.data.api
-
-import com.bignerdranch.android.translator.model.data.DataModel
-import kotlinx.coroutines.Deferred
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ApiService {
-
-    @GET("words/search")
-    fun searchAsync(@Query("search") wordToSearch: String): Deferred<List<DataModel>>
-}
