--- app/src/main/java/com/bignerdranch/android/translator/model/datasource/DataSource.kt
@@ -1,7 +1,0 @@
-package com.bignerdranch.android.translator.model.datasource
-
-
-interface DataSource<T> {
-
-    suspend fun getData(word: String): T
-}
