--- app/src/main/java/com/bignerdranch/android/translator/model/datasource/RetrofitImplementation.kt
@@ -1,44 +1,0 @@
-package com.bignerdranch.android.translator.model.datasource
-
-
-import com.bignerdranch.android.translator.model.data.DataModel
-import com.bignerdranch.android.translator.model.data.api.ApiService
-import com.bignerdranch.android.translator.model.data.api.BaseInterceptor
-import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
-
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class RetrofitImplementation : DataSource<List<DataModel>> {
-
-    override suspend fun getData(word: String): List<DataModel> {
-        return getService(BaseInterceptor.interceptor).searchAsync(word).await()
-    }
-
-    private fun getService(interceptor: Interceptor): ApiService {
-        return createRetrofit(interceptor).create(ApiService::class.java)
-    }
-
-    private fun createRetrofit(interceptor: Interceptor): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(BASE_URL_LOCATIONS)
-            .addConverterFactory(GsonConverterFactory.create())
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
-            .client(createOkHttpClient(interceptor))
-            .build()
-    }
-
-    private fun createOkHttpClient(interceptor: Interceptor): OkHttpClient {
-        val httpClient = OkHttpClient.Builder()
-        httpClient.addInterceptor(interceptor)
-        httpClient.addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-        return httpClient.build()
-    }
-
-    companion object {
-        private const val BASE_URL_LOCATIONS = "https://dictionary.skyeng.ru/api/public/v1/"
-    }
-}
