--- app/src/main/java/com/bignerdranch/android/translator/model/repository/Repository.kt
@@ -1,6 +1,0 @@
-package com.bignerdranch.android.translator.model.repository
-
-
-interface Repository<T> {
-    suspend fun getData(word: String): T
-}
