--- app/src/main/java/com/bignerdranch/android/translator/model/repository/RepositoryImplementation.kt
@@ -1,12 +1,0 @@
-package com.bignerdranch.android.translator.model.repository
-
-import com.bignerdranch.android.translator.model.data.DataModel
-import com.bignerdranch.android.translator.model.datasource.DataSource
-
-class RepositoryImplementation(private val dataSource: DataSource<List<DataModel>>) :
-    Repository<List<DataModel>> {
-
-    override suspend fun getData(word: String): List<DataModel> {
-        return dataSource.getData(word)
-    }
-}
