--- app/src/main/java/com/bignerdranch/android/translator/model/repository/RepositoryLocal.kt
@@ -1,7 +1,0 @@
-package com.bignerdranch.android.translator.model.repository
-
-import com.bignerdranch.android.translator.model.data.AppState
-
-interface RepositoryLocal<T>:Repository<T> {
-    suspend fun saveToDB(appState: AppState)
-}
