--- app/src/main/java/com/bignerdranch/android/translator/room/HistoryDao.kt
@@ -1,24 +1,0 @@
-package com.bignerdranch.android.translator.room
-
-import androidx.room.*
-
-@Dao
-interface HistoryDao {
-    @Query("SELECT * FROM HistoryEntity")
-    suspend fun all(): List<HistoryEntity>
-
-    @Query("SELECT * FROM HistoryEntity WHERE word LIKE :word")
-    suspend fun getDataByWord(word: String): HistoryEntity
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insert(entity: HistoryEntity)
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insertAll(entities: List<HistoryEntity>)
-
-    @Update
-    suspend fun update(entity: HistoryEntity)
-
-    @Delete
-    suspend fun delete(entity: HistoryEntity)
-}
