--- app/src/main/java/com/bignerdranch/android/translator/room/HistoryDataBase.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.android.translator.room
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-@Database(entities = arrayOf(HistoryEntity::class), version = 1, exportSchema = false)
-abstract class HistoryDataBase :RoomDatabase(){
-    abstract fun historyDao():HistoryDao
-}
