--- app/src/main/java/com/bignerdranch/android/translator/room/HistoryEntity.kt
@@ -1,15 +1,0 @@
-package com.bignerdranch.android.translator.room
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.Index
-import androidx.room.PrimaryKey
-
-@Entity(indices = arrayOf(Index(value = arrayOf("word"), unique = true)))
-class HistoryEntity(
-    @field:PrimaryKey
-    @field:ColumnInfo(name = "word")
-    var word: String, @field:ColumnInfo(name = "description")
-    var description: String?
-)
-
