+++ repository/src/main/java/com/bignerdranch/android/repository/DataSource.kt
@@ -1,0 +1,6 @@
+package com.bignerdranch.android.repository
+
+interface DataSource<T> {
+
+    suspend fun getData(word: String): T
+}
