+++ repository/src/main/java/com/bignerdranch/android/repository/DataSourceLocal.kt
@@ -1,0 +1,9 @@
+package com.bignerdranch.android.repository
+
+import com.bignerdranch.android.model.data.AppState
+
+
+interface DataSourceLocal<T> : DataSource<T> {
+
+    suspend fun saveToDB(appState: AppState)
+}
