+++ repository/src/main/java/com/bignerdranch/android/repository/Repository.kt
@@ -1,0 +1,6 @@
+package com.bignerdranch.android.repository
+
+interface Repository<T> {
+
+    suspend fun getData(word: String): T
+}
