+++ repository/src/main/java/com/bignerdranch/android/repository/RepositoryImplementation.kt
@@ -1,0 +1,12 @@
+package com.bignerdranch.android.repository
+
+import com.bignerdranch.android.model.data.DataModel
+
+
+class RepositoryImplementation(private val dataSource: DataSource<List<DataModel>>) :
+    Repository<List<DataModel>> {
+
+    override suspend fun getData(word: String): List<DataModel> {
+        return dataSource.getData(word)
+    }
+}
