+++ repository/src/main/java/com/bignerdranch/android/repository/RepositoryImplementationLocal.kt
@@ -1,0 +1,17 @@
+package com.bignerdranch.android.repository
+
+import com.bignerdranch.android.model.data.AppState
+import com.bignerdranch.android.model.data.DataModel
+
+
+class RepositoryImplementationLocal(private val dataSource: DataSourceLocal<List<DataModel>>) :
+    RepositoryLocal<List<DataModel>> {
+
+    override suspend fun getData(word: String): List<DataModel> {
+        return dataSource.getData(word)
+    }
+
+    override suspend fun saveToDB(appState: AppState) {
+        dataSource.saveToDB(appState)
+    }
+}
