+++ repository/src/main/java/com/bignerdranch/android/repository/RepositoryLocal.kt
@@ -1,0 +1,9 @@
+package com.bignerdranch.android.repository
+
+import com.bignerdranch.android.model.data.AppState
+
+
+interface RepositoryLocal<T> : Repository<T> {
+
+    suspend fun saveToDB(appState: AppState)
+}
