+++ repository/src/main/java/com/bignerdranch/android/repository/RoomDataBaseImplementation.kt
@@ -1,0 +1,19 @@
+package com.bignerdranch.android.repository
+
+import com.bignerdranch.android.model.data.AppState
+import com.bignerdranch.android.model.data.DataModel
+import com.bignerdranch.android.repository.room.HistoryDao
+
+class RoomDataBaseImplementation(private val historyDao: HistoryDao) :
+    DataSourceLocal<List<DataModel>> {
+
+    override suspend fun getData(word: String): List<DataModel> {
+        return mapHistoryEntityToSearchResult(historyDao.all())
+    }
+
+    override suspend fun saveToDB(appState: AppState) {
+        convertDataModelSuccessToEntity(appState)?.let {
+            historyDao.insert(it)
+        }
+    }
+}
