+++ repository/src/main/java/com/bignerdranch/android/repository/SearchResultParser.kt
@@ -1,0 +1,30 @@
+package com.bignerdranch.android.repository
+
+
+import com.bignerdranch.android.model.data.AppState
+import com.bignerdranch.android.model.data.DataModel
+import com.bignerdranch.android.repository.room.HistoryEntity
+
+fun mapHistoryEntityToSearchResult(list: List<HistoryEntity>): List<DataModel> {
+    val searchResult = ArrayList<DataModel>()
+    if (!list.isNullOrEmpty()) {
+        for (entity in list) {
+            searchResult.add(DataModel(entity.word, null))
+        }
+    }
+    return searchResult
+}
+
+fun convertDataModelSuccessToEntity(appState: AppState): HistoryEntity? {
+    return when (appState) {
+        is AppState.Success -> {
+            val searchResult = appState.data
+            if (searchResult.isNullOrEmpty() || searchResult[0].text.isNullOrEmpty()) {
+                null
+            } else {
+                HistoryEntity(searchResult[0].text!!, null)
+            }
+        }
+        else -> null
+    }
+}
