+++ repository/src/main/java/com/bignerdranch/android/repository/api/ApiService.kt
@@ -1,0 +1,13 @@
+package com.bignerdranch.android.repository.api
+
+
+import com.bignerdranch.android.model.data.DataModel
+import kotlinx.coroutines.Deferred
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ApiService {
+
+    @GET("words/search")
+    fun searchAsync(@Query("search") wordToSearch: String): Deferred<List<DataModel>>
+}
