+++ repository/src/main/java/com/bignerdranch/android/repository/room/HistoryDataBase.kt
@@ -1,0 +1,10 @@
+package com.bignerdranch.android.repository.room
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(entities = arrayOf(HistoryEntity::class), version = 1, exportSchema = false)
+abstract class HistoryDataBase : RoomDatabase() {
+
+    abstract fun historyDao(): HistoryDao
+}
