+++ repository/src/main/java/com/bignerdranch/android/repository/room/HistoryEntity.kt
@@ -1,0 +1,14 @@
+package com.bignerdranch.android.repository.room
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
+
+@Entity(indices = arrayOf(Index(value = arrayOf("word"), unique = true)))
+class HistoryEntity(
+    @field:PrimaryKey
+    @field:ColumnInfo(name = "word")
+    var word: String, @field:ColumnInfo(name = "description")
+    var description: String?
+)
