+++ settings.gradle
@@ -2,3 +2,4 @@ include ':app'
 
 include ':utils'
 include ':model'
+include ':repository'
